BIGTOP-2414. Rename Tachyon to Alluxio
