+++ MAINTAINERS.txt
@@ -1,3 +1,4 @@
+alluxio: jay vyas, huamin chen
 apex: chinmay <chinmay@apache.org>, aniruddha <aniruddha@apache.org>
 bigpetstore: jay vyas, rj nowling
 CI infra: rvs
@@ -21,7 +22,6 @@ qfs: fsareshwala <fsareshwala@quantcast.com>
 smoke-tests: jay vyas, david capwell
 spark: jay vyas, youngwoo kim
 sqoop: sean mackrory, youngwoo kim
-tachyon: jay vyas, huamin chen
 tajo: yeongeon kim
 test-artifacts and test-execution repos:
 tez: evans ye, oflebbe
