+++ bigtop-ci/jenkins/jobsCreator.groovy
@@ -19,7 +19,7 @@
 def bigtopComponents = ["bigtop-groovy", "bigtop-jsvc", "bigtop-tomcat", "bigtop-utils", 
                         "zookeeper", "hadoop", "hbase", "hive", "pig", "crunch", "datafu", 
                         "flume", "giraph", "ignite-hadoop", "mahout", "oozie", "phoenix",
-                        "solr", "spark", "sqoop", "tachyon", "whirr"]
+                        "solr", "spark", "sqoop", "alluxio", "whirr"]
 // FIXME: it would be nice to extract the following from some static configuration file
 def targetOS = ["fedora-20",  "opensuse-12.3",  "ubuntu-14.04"]
 def gitUrl = "https://git-wip-us.apache.org/repos/asf/bigtop.git"
