+++ bigtop-deploy/puppet/modules/alluxio/templates/log4j.properties
@@ -13,9 +13,9 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 # May get overridden by System Property
-tachyon.logger.type=Console
+alluxio.logger.type=Console
 
-log4j.rootLogger=INFO, ${tachyon.logger.type}
+log4j.rootLogger=INFO, ${alluxio.logger.type}
 
 log4j.appender.Console=org.apache.log4j.ConsoleAppender
 log4j.appender.Console.Target=System.out
@@ -23,8 +23,8 @@ log4j.appender.Console.layout=org.apache.log4j.PatternLayout
 log4j.appender.Console.layout.ConversionPattern=%d{ISO8601} %-5p %c{1} (%F:%M) - %m%n
 
 # Appender for Master
-log4j.appender.MASTER_LOGGER=tachyon.Log4jFileAppender
-log4j.appender.MASTER_LOGGER.File=/var/log/tachyon/master.log
+log4j.appender.MASTER_LOGGER=alluxio.Log4jFileAppender
+log4j.appender.MASTER_LOGGER.File=/var/log/alluxio/master.log
 
 log4j.appender.MASTER_LOGGER.MaxFileSize=10
 log4j.appender.MASTER_LOGGER.MaxBackupIndex=100
@@ -34,8 +34,8 @@ log4j.appender.MASTER_LOGGER.layout.ConversionPattern=%d{ISO8601} %-5p %c{2} (%F
 #log4j.appender.MASTER_LOGGER.layout.ConversionPattern=%d{ISO8601} %-5p %c{2} (%F:%M(%L)) - %m%n
 
 # Appender for Workers
-log4j.appender.WORKER_LOGGER=tachyon.Log4jFileAppender
-log4j.appender.WORKER_LOGGER.File=/var/log/tachyon/worker.log
+log4j.appender.WORKER_LOGGER=alluxio.Log4jFileAppender
+log4j.appender.WORKER_LOGGER.File=/var/log/alluxio/worker.log
 
 log4j.appender.WORKER_LOGGER.MaxFileSize=10
 log4j.appender.WORKER_LOGGER.MaxBackupIndex=100
@@ -45,8 +45,8 @@ log4j.appender.WORKER_LOGGER.layout.ConversionPattern=%d{ISO8601} %-5p %c{2} (%F
 #log4j.appender.WORKER_LOGGER.layout.ConversionPattern=%d{ISO8601} %-5p %c{2} (%F:%M(%L)) - %m%n
 
 # Appender for User
-log4j.appender.USER_LOGGER=tachyon.Log4jFileAppender
-log4j.appender.USER_LOGGER.File=/var/log/tachyon/logs/user.log
+log4j.appender.USER_LOGGER=alluxio.Log4jFileAppender
+log4j.appender.USER_LOGGER.File=/var/log/alluxio/logs/user.log
 log4j.appender.USER_LOGGER.MaxFileSize=10
 log4j.appender.USER_LOGGER.MaxBackupIndex=10
 log4j.appender.USER_LOGGER.DeletionPercentage=20
