+++ bigtop-packages/src/common/alluxio/alluxio-worker.svc
@@ -14,14 +14,14 @@
 # limitations under the License.
 
 TYPE="worker"
-DAEMON="tachyon-${TYPE}"
-DESC="Tachyon ${TYPE}"
-EXEC_PATH="/usr/lib/tachyon/bin/tachyon-start.sh"
-WORKING_DIR="/var/lib/tachyon"
+DAEMON="alluxio-${TYPE}"
+DESC="Alluxio ${TYPE}"
+EXEC_PATH="/usr/lib/alluxio/bin/alluxio-start.sh"
+WORKING_DIR="/var/lib/alluxio"
 DAEMON_FLAGS="worker Mount"
-KEYWORD="tachyon.worker.TachyonWorker"
-CONF_DIR="/etc/tachyon/conf"
-PIDFILE="/var/run/tachyon/${DAEMON}.pid"
+KEYWORD="alluxio.worker.AlluxioWorker"
+CONF_DIR="/etc/alluxio/conf"
+PIDFILE="/var/run/alluxio/${DAEMON}.pid"
 
 generate_start() {
 
@@ -37,10 +37,10 @@ start() {
         exit 0
     fi
 
-    LOG_FILE=/var/log/tachyon/${DAEMON}.out
+    LOG_FILE=/var/log/alluxio/${DAEMON}.out
 
-    if [ -f $CONF_DIR/tachyon-env.sh ]; then
-        . $CONF_DIR/tachyon-env.sh
+    if [ -f $CONF_DIR/alluxio-env.sh ]; then
+        . $CONF_DIR/alluxio-env.sh
     fi
 
     /bin/bash -c "nohup nice -n 0 \
@@ -65,7 +65,7 @@ generate_stop() {
 cat <<'__EOT__'
 stop() {
     log_success_msg "Stopping $DESC (${DAEMON}): "
-    /usr/lib/tachyon/bin/tachyon-stop.sh worker 2>&1 > /dev/null
+    /usr/lib/alluxio/bin/alluxio-stop.sh worker 2>&1 > /dev/null
     RETVAL=$?
 
     [ $RETVAL -eq $RETVAL_SUCCESS ] && rm -f $LOCKFILE $PIDFILE
