+++ bigtop-packages/src/common/alluxio/init.d.tmpl
@@ -29,7 +29,7 @@
 #     EXEC_PATH="path to the upstream daemon management script"
 #     SVC_USER="user to run this service as"
 #     DAEMON_FLAGS="flags to be passed to the $EXEC_PATH"
-#     KEYWORD="keyword to determine the Tachyon daemon"
+#     KEYWORD="keyword to determine the Alluxio daemon"
 #     WORKING_DIR="working directory to cd into before starting the daemon"
 #     CONF_DIR="path to the configuration directory"
 #     PIDFILE="file holding a PID of the running daemon"
