+++ bigtop-packages/src/common/alluxio/install_alluxio.sh
@@ -21,12 +21,12 @@ usage() {
   echo "
 usage: $0 <options>
   Required not-so-options:
-     --build-dir=DIR             path to Tachyon dist.dir
+     --build-dir=DIR             path to Alluxio dist.dir
      --prefix=PREFIX             path to install into
 
   Optional options:
      --bin-dir=DIR               path to install bin
-     --data-dir=DIR              path to install local Tachyon data
+     --data-dir=DIR              path to install local Alluxio data
      ... [ see source for more similar options ]
   "
   exit 1
@@ -89,7 +89,7 @@ for var in PREFIX BUILD_DIR ; do
   fi
 done
 
-LIB_DIR=${LIB_DIR:-/usr/lib/tachyon}
+LIB_DIR=${LIB_DIR:-/usr/lib/alluxio}
 LIBEXEC_DIR=${INSTALLED_LIB_DIR:-/usr/libexec}
 BIN_DIR=${BIN_DIR:-/usr/bin}
 
@@ -99,55 +99,54 @@ install -d -m 0755 $PREFIX/$LIB_DIR/libexec
 install -d -m 0755 $PREFIX/$LIB_DIR/lib
 install -d -m 0755 $PREFIX/$LIB_DIR/share
 install -d -m 0755 $PREFIX/$DATA_DIR
-install -d -m 0755 $PREFIX/$DATA_DIR/tachyon
+install -d -m 0755 $PREFIX/$DATA_DIR/alluxio
 install -d -m 0755 $PREFIX/etc
-install -d -m 0755 $PREFIX/etc/tachyon
-install -d -m 0755 $PREFIX/etc/tachyon/conf
-install -d -m 0755 $PREFIX/$VAR_DIR/log/tachyon
-install -d -m 0755 $PREFIX/$VAR_DIR/lib/tachyon/journal
-install -d -m 0755 $PREFIX/$VAR_DIR/lib/tachyon/core/src/main/webapp
-ln -s $VAR_DIR/log/tachyon $PREFIX/$VAR_DIR/lib/tachyon/logs
-install -d -m 0755 $PREFIX/$VAR_DIR/run/tachyon
+install -d -m 0755 $PREFIX/etc/alluxio
+install -d -m 0755 $PREFIX/etc/alluxio/conf
+install -d -m 0755 $PREFIX/$VAR_DIR/log/alluxio
+install -d -m 0755 $PREFIX/$VAR_DIR/lib/alluxio/journal
+install -d -m 0755 $PREFIX/$VAR_DIR/lib/alluxio/core/server/src/main/webapp
+ln -s $VAR_DIR/log/alluxio $PREFIX/$VAR_DIR/lib/alluxio/logs
+install -d -m 0755 $PREFIX/$VAR_DIR/run/alluxio
 
-#cp -ra ${BUILD_DIR}/lib/* $PREFIX/${LIB_DIR}/lib/
-cp client/target/tachyon-client*.jar core/target/tachyon*.jar $PREFIX/$LIB_DIR
+cp assembly/target/alluxio*dependencies.jar core/client/target/alluxio*dependencies.jar $PREFIX/$LIB_DIR
 cp -a bin/* $PREFIX/${LIB_DIR}/bin
 cp -a libexec/* $PREFIX/${LIB_DIR}/libexec
-cp -rf core/src/main/webapp $PREFIX/$VAR_DIR/lib/tachyon/core/src/main
+cp -rf core/server/src/main/webapp $PREFIX/$VAR_DIR/lib/alluxio/core/server/src/main
 
 # Copy in the configuration files
-install -m 0644 conf/log4j.properties conf/workers $PREFIX/etc/tachyon/conf
-cp conf/tachyon-env.sh.template $PREFIX/etc/tachyon/conf/tachyon-env.sh
+install -m 0644 conf/log4j.properties $PREFIX/etc/alluxio/conf
+cp conf/alluxio-env.sh.template $PREFIX/etc/alluxio/conf/alluxio-env.sh
 
-# Copy in the /usr/bin/tachyon wrapper
+# Copy in the /usr/bin/alluxio wrapper
 install -d -m 0755 $PREFIX/$BIN_DIR
 
-# Copy in tachyon deploy scripts
+# Copy in alluxio deploy scripts
 cp -rf deploy $PREFIX/$LIB_DIR/share
 
 # Prefix is correct at time of install,
 # but we dont want to escape it before that point.
-cat > $PREFIX/$BIN_DIR/tachyon <<EOF
+cat > $PREFIX/$BIN_DIR/alluxio <<EOF
 #!/bin/bash
 
 # Autodetect JAVA_HOME if not defined
 . /usr/lib/bigtop-utils/bigtop-detect-javahome
 # Lib dir => ${LIB_DIR}
 #!/usr/bin/env bash
-exec ${LIB_DIR}/bin/tachyon "\$@"
+exec ${LIB_DIR}/bin/alluxio "\$@"
 EOF
-chmod 755 $PREFIX/$BIN_DIR/tachyon
+chmod 755 $PREFIX/$BIN_DIR/alluxio
 
-cat >$PREFIX/$LIB_DIR/libexec/tachyon-layout.sh <<EOF
+cat >$PREFIX/$LIB_DIR/libexec/alluxio-layout.sh <<EOF
 #!/usr/bin/env bash
 
-export TACHYON_SYSTEM_INSTALLATION="TRUE"
-export TACHYON_PREFIX="$LIB_DIR"
-export TACHYON_HOME="/var/lib/tachyon"
-export TACHYON_CONF_DIR="/etc/tachyon/conf"
-export TACHYON_LOGS_DIR="/var/log/tachyon"
-export TACHYON_DATA_DIR="/var/run/tachyon"
-export TACHYON_JAR="\`find $LIB_DIR/ -name tachyon*dependencies.jar|grep -v client\`"
+export ALLUXIO_SYSTEM_INSTALLATION="TRUE"
+export ALLUXIO_PREFIX="$LIB_DIR"
+export ALLUXIO_HOME="/var/lib/alluxio"
+export ALLUXIO_CONF_DIR="/etc/alluxio/conf"
+export ALLUXIO_LOGS_DIR="/var/log/alluxio"
+export ALLUXIO_DATA_DIR="/var/run/alluxio"
+export ALLUXIO_JARS="\`find $LIB_DIR/ -name alluxio*dependencies.jar|grep -v client\`"
 
 # find JAVA_HOME
 . /usr/lib/bigtop-utils/bigtop-detect-javahome
