+++ bigtop-packages/src/deb/alluxio/alluxio.dirs
@@ -1,0 +1,4 @@
+/etc/alluxio/conf
+/usr/lib/alluxio
+/usr/bin
+/var/run/alluxio
