+++ bigtop-packages/src/deb/alluxio/alluxio.install
@@ -1,0 +1,3 @@
+/etc/alluxio/conf
+/usr/lib/alluxio
+/usr/bin/alluxio
