+++ bigtop-packages/src/deb/alluxio/control
@@ -13,25 +13,25 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-Source: tachyon-tfs
+Source: alluxio
 Section: misc
 Priority: extra
 Maintainer: Huamin Chen <hchen@redhat.com>
 Build-Depends: debhelper (>= 7.0.50~)
 Standards-Version: 3.9.4
-Homepage: http://tachyon-project.org
+Homepage: http://www.alluxio.org
 
-Package:  tachyon-tfs
+Package:  alluxio
 Architecture: all
 Depends: bigtop-utils (>= 0.7)
 Description: Reliable file sharing at memory speed across cluster frameworks
- Tachyon is a fault tolerant distributed file system
+ Alluxio is a fault tolerant distributed file system
  enabling reliable file sharing at memory-speed
  across cluster frameworks, such as Spark and MapReduce.
  It achieves high performance by leveraging lineage
  information and using memory aggressively.
- Tachyon caches working set files in memory, and
+ Alluxio caches working set files in memory, and
  enables different jobs/queries and frameworks to
- access cached files at memory speed. Thus, Tachyon
+ access cached files at memory speed. Thus, Alluxio
  avoids going to disk to load data-sets that
  are frequently read.
