+++ bigtop-packages/src/deb/alluxio/copyright
@@ -1,6 +1,6 @@
 Format: http://dep.debian.net/deps/dep5
-Source: http://tachyon-project.org/
-Upstream-Name: Tachyon Project
+Source: http://www.alluxio.org/
+Upstream-Name: Alluxio Project
 
 Files: *
 Copyright: 2010-2011, The Apache Software Foundation
