--- bigtop-packages/src/deb/tachyon/patch0-exclude-debian-subdir-from-license-check.diff
@@ -1,10 +1,0 @@
---- tachyon-0.6.0/pom.xml.orig	2015-02-11 06:23:19.000000000 +0100
-+++ tachyon-0.6.0/pom.xml	2015-04-01 21:52:11.672000000 +0200
-@@ -302,6 +302,7 @@
-             <exclude>**/docs/**/*</exclude>
-             <exclude>LICENSE</exclude>
-             <exclude>NOTICE</exclude>
-+            <exclude>debian/**</exclude>
-           </excludes>
-           <mapping>
-             <java>SLASHSTAR_STYLE</java>
