--- bigtop-packages/src/deb/tachyon/tachyon-tfs.dirs
@@ -1,4 +1,0 @@
-/etc/tachyon/conf
-/usr/lib/tachyon
-/usr/bin
-/var/run/tachyon
