--- bigtop-packages/src/deb/tachyon/tachyon-tfs.install
@@ -1,3 +1,0 @@
-/etc/tachyon/conf
-/usr/lib/tachyon
-/usr/bin/tachyon
