+++ bigtop-tests/smoke-tests/alluxio/datafile
@@ -1,0 +1,1 @@
+Test file to see if Hadoop is able to read this file from Alluxio.
