--- bigtop-tests/smoke-tests/tachyon/datafile
@@ -1,1 +1,0 @@
-Test file to see if Hadoop is able to read this file from Tachyon.
